+++ src/pages/index.js
@@ -3,7 +3,6 @@ import NormalModePage from './NormalModePage';
 import SplashPage from './SplashPage';
 import MainPage from './MainPage';
 import LoginPage from './LoginPage';
-import SavePage from './SavePage';
 import FeedbackPage from './FeedbackPage';
 import JoinPage from './JoinPage';
 
@@ -14,6 +13,5 @@ export {
   SplashPage,
   MainPage,
   LoginPage,
-  SavePage,
   FeedbackPage,
 };
